Change client interface to Reactive Streams
Rather than offering a callback mechanism to serve the received
messages, this commit introduces a Reactive Streams based API for this
component.

Internally, it uses RxJava 2.0 to provide the Reactive implementaion.

Class BSB1MessagePublisher is responsible for managing the subscriptions
and publishing the messages. TODO: proper concurrent behavior,
multiple subscribers and honoring the requested number of items.

Netty's BSB1ClientHandler is now initialized with a reference to a
BSB1MessagePublisher and uses it to publish the messages when decoded.
If no subscriber is subscribed, messages are still decoded from the
TCP/IP connection but not published.

The new client interface to interacting with this component is similar
to:

BSB1Client client = BSB1Client
    .instance()
    .host("localhost")
    .port(30003)
    .subscribe(new Subscriber<BSB1CSVMessage>() {
        @Override
        public void onSubscribe(Subscription s) {
          s.request(Long.MAX_VALUE);
        }

        @Override
        public void onNext(BSB1CSVMessage message) {
          System.out.println(message);
        }

        @Override
        public void onError(Throwable t) {
          t.printStackTrace();
        }

        @Override
        public void onComplete() {}
    })
    .get();

client.start();

or use a ReactiveStreams provider for client simplicity.