Use FlowableProcessor to generate BSB1CSVMessage
Previous version used Flowable.create and FlowableEmitter to generate
the messages. But it was inadequate: those are used for cold Flowables.

Instead, a FlowableProcessor needs to be used to turn the async events
that originated (in this case from Netty's async
MessageToMessageDecoder) into the reactive Flowable<BSB1CSVMessage>.
More precisely, a PublishProcessor is used. Code becomes also simpler to
use.

A call to onComplete() is added if the channel is closed. This allows
for .count() methods to work on the stream, though future versions of
the client should never stop processing events, even if the client is
stopped (it should try to re-connect again).