Implement producer/consumer separation w Disruptor
Before this commit, producer and consumer speeds were linked. In the
business logic of this program, they should operate completely
independently and if the consumer is slower, it makes sense some
producer messages may be discarded.

This commit uses LMAX's Disruptor to provide a clear separation between
both. This also helps reduce to the bare minimum the time the Netty
threads are used processing the request. Now, as soon as a String is
decoded from Netty, it is pushed to the Disruptor and Netty activity
finishes.

Then Disruptor buffers a limited amount of BSB1CSVMessage and the client
now reads the messages from the Disruptor. Indeed, now Netty's client
handler does not need to now about RxJava or the consumer of the
messages in general.

If the Disruptor's buffer is full (consumer is slower than producer)
received messages start to be discarded.

Simple initial tests prove this works very well, and both Disruptor and
RxJava2 are very fast. Disruptor is faster than consuming the messages
via RxJava2, which was expected.

All in all, this commit introduces a significant improvement to the
code, performance and backpressure strategy.

BSB1CSVMessage was changed to a mutable form, since Disruptor requires
events to be pre-allocated (which in turns require a factory method to
instantiate and reuse them). For our case, this should also slightly
reduce GC.